+++ node/src/Worker.ts
@@ -293,6 +293,16 @@ export class WorkerImpl<WorkerAppData extends AppData = AppData>
 			}
 		});
 
+		// NOTE: Avoid "Possible EventEmitter memory leak detected" Node warning.
+		const processListenerCount = process.getMaxListeners();
+
+		if (processListenerCount >= 10) {
+			process.setMaxListeners(processListenerCount + 2);
+		}
+
+		process.once('SIGINT', this.onSignal);
+		process.once('SIGTERM', this.onSignal);
+
 		this.handleListenerError();
 	}
 
@@ -346,6 +356,9 @@ export class WorkerImpl<WorkerAppData extends AppData = AppData>
 		logger.debug('close()');
 
 		this.#closed = true;
+
+		process.removeListener('SIGINT', this.onSignal);
+		process.removeListener('SIGTERM', this.onSignal);
 
 		// Close every Router.
 		for (const router of this.#routers) {
@@ -593,6 +606,15 @@ export class WorkerImpl<WorkerAppData extends AppData = AppData>
 			);
 		});
 	}
+
+	// NOTE: Arrow method on purpose.
+	private onSignal = (signal: 'SIGINT' | 'SIGTERM'): void => {
+		logger.debug(
+			`signal received, closing the worker process [pid:${this.#pid}, signal:${signal}]`
+		);
+
+		this.close();
+	};
 }
 
 function parseWorkerDumpResponse(binary: FbsWorker.DumpResponse): WorkerDump {